% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{foo}
\alias{foo}
\alias{foo,{-method}
\title{Foo (generic)}
\usage{
foo(x, ...)

\S4method{foo}{`{`}(x, ...)
}
\arguments{
\item{x}{\strong{Signature argument}.
Object containing expression information. Typically, a function or an
"curly expression/call".}

\item{x}{\code{\link{{}}}.}
}
\value{
\code{\link{expression}}.
}
\description{
Blabla.
}
\section{Methods (by class)}{
\itemize{
\item \code{{}: 
}}

